feat: more frontend and state management
