+++ src/components/FormComponents/Error.jsx
@@ -1,0 +1,11 @@
+import { FormErrorMessage } from "@chakra-ui/react";
+import { useField } from "react-final-form";
+
+const Error = ({ name }) => {
+  const {
+    meta: { error },
+  } = useField(name, { subscription: { error: true } });
+  return <FormErrorMessage>{error}</FormErrorMessage>;
+};
+
+export default Error;
