+++ src/components/FormComponents/Multiselect.jsx
@@ -1,0 +1,33 @@
+import { FormLabel, HStack, Stack } from "@chakra-ui/react";
+import CheckboxArrayControl from "./CheckboxArrayControl";
+import Control from "./Control";
+import Error from "./Error";
+
+const Multiselect = ({ name, label, entries }) => {
+  return (
+    <Control name={name} my={4}>
+      <HStack>
+        <FormLabel
+          htmlFor={name}
+          m={0}
+          fontSize="xl"
+          fontWeight="bold"
+          color="#8C8C8C"
+        >
+          {label}
+        </FormLabel>
+        <Error name={name} />
+      </HStack>
+
+      <Stack mt={1} spacing={1}>
+        {entries.map((e, index) => (
+          <CheckboxArrayControl key={index} name={name} value={e}>
+            {e}
+          </CheckboxArrayControl>
+        ))}
+      </Stack>
+    </Control>
+  );
+};
+
+export default Multiselect;
