+++ src/components/FormComponents/TextareaControl.jsx
@@ -1,0 +1,22 @@
+import { FormLabel, HStack, Textarea } from "@chakra-ui/react";
+import { Field } from "react-final-form";
+import Control from "./Control";
+import Error from "./Error";
+
+const AdaptedTextarea = ({ input, meta, ...rest }) => (
+  <Textarea {...input} {...rest} isInvalid={meta.error && meta.touched} />
+);
+
+const TextareaControl = ({ name, label, ...props }) => (
+  <Control name={name}>
+    <HStack>
+      <FormLabel htmlFor={name} m={0} color="#6D6E70">
+        {label}
+      </FormLabel>
+      <Error name={name} />
+    </HStack>
+    <Field name={name} component={AdaptedTextarea} id={name} {...props} />
+  </Control>
+);
+
+export default TextareaControl;
